Move code from __init__.py to own module
