+++ rcon/cli.py
@@ -5,7 +5,7 @@ from datetime import datetime, timedelta
 
 import click
 
-import rcon.expiring_vips
+import rcon.expiring_vips.service
 from rcon import auto_settings, broadcast, game_logs, routines, stats_loop
 from rcon.cache_utils import RedisCached, get_redis_pool
 from rcon.extended_commands import Rcon
@@ -80,7 +80,7 @@ def run_routines():
 
 @cli.command(name="expiring_vips")
 def run_expiring_vips():
-    rcon.expiring_vips.run()
+    rcon.expiring_vips.service.run()
 
 
 @cli.command(name="noleaders")
